requirements.txt
# Image Processing Service - Python Dependencies
# Version: 1.0.0

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.19  # File upload support

# Data Validation
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Database
supabase==2.12.0
postgrest==0.19.3

# Image Processing
opencv-python==4.10.0.84
pillow==11.0.0
numpy==2.2.1

# OCR
pytesseract==0.3.14
pdfplumber==0.11.4

# AI/LLM
openai==1.59.5

# Fuzzy Matching
rapidfuzz==3.10.1

# QR Code & PDF Generation
qrcode[pil]==8.0
reportlab==4.2.5

# Authentication
pyjwt==2.10.1
cryptography==44.0.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1  # Async HTTP client
aiofiles==24.1.0  # Async file operations
tenacity==9.0.0  # Retry logic

# Monitoring & Logging
structlog==24.4.0

# Testing (dev only)
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx[test]==0.28.1  # Test client for FastAPI

# Code Quality (dev only)
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Hashing
hashlib-additional==1.0  # SHA256 for file deduplication
